Delete duplicate FloatRasterVAT.
